+++ src/services/axios/get/getProposal/getProposal.tsx
@@ -26,6 +26,8 @@ import {
   USE_LOCAL_DATA,
   DETAILS,
   TYPE_CONTINUUM,
+  TYPE_CONTINUUM_SPECTRAL,
+  TYPE_CONTINUUM_SPECTRAL_LONG,
   TYPE_ZOOM,
   VEL_TYPES,
   VEL_UNITS,
@@ -448,7 +450,9 @@ const getLinked = (
 
 // This is here as there is inconsistent representation of spectral and spectral line.
 const typeCheck = (inType: string | undefined): any => {
-  return inType === TYPE_ZOOM_LONG ? TYPE_ZOOM : inType;
+  if (inType === TYPE_ZOOM_LONG) return TYPE_ZOOM;
+  if (inType === TYPE_CONTINUUM_SPECTRAL_LONG) return TYPE_CONTINUUM_SPECTRAL;
+  return inType;
 };
 
 const getObservations = (
@@ -526,11 +530,11 @@ const getObservations = (
       ),
       linked: getLinked(inValue[i], inResults),
       continuumBandwidth:
-        type === TYPE_CONTINUUM || type === TYPE_PST
+        type === TYPE_CONTINUUM || type === TYPE_PST || type === TYPE_CONTINUUM_SPECTRAL
           ? (inValue[i].observation_type_details?.bandwidth?.value ?? null)
           : null,
       continuumBandwidthUnits:
-        type === TYPE_CONTINUUM || type === TYPE_PST
+        type === TYPE_CONTINUUM || type === TYPE_PST || type === TYPE_CONTINUUM_SPECTRAL
           ? getFrequencyAndBandwidthUnits(
               inValue[i]?.observation_type_details?.bandwidth?.unit ?? null,
               observingBand
@@ -717,7 +721,8 @@ const getTargetObservation = (
   }
   for (const result of inResults) {
     const resultObsType = getResultObsType(result, inObservationSets);
-    const isContinuum = resultObsType === TYPE_CONTINUUM;
+    const isContinuum =
+      resultObsType === TYPE_CONTINUUM || resultObsType === TYPE_CONTINUUM_SPECTRAL_LONG;
     const isPST = resultObsType === TYPE_PST;
     const isSensitivity = result.result?.supplied_type === 'sensitivity';